[UWP] Frames logic are now in their corresponding view model. Frames now have their corresponding page
