+++ MiraiNotes.UWP/MiraiNotes.UWP.csproj
@@ -101,6 +101,8 @@
     <Compile Include="Converters\BooleanToVisibilityConverter.cs" />
     <Compile Include="Converters\NavigationViewSelectionChangedEventArgsConverter.cs" />
     <Compile Include="Converters\TestConverter.cs" />
+    <Compile Include="Design\DesignGoogleTaskListService.cs" />
+    <Compile Include="Design\DesignGoogleTaskService.cs" />
     <Compile Include="Extensions\FocusExtension.cs" />
     <Compile Include="Handlers\AuthorizationHandler.cs" />
     <Compile Include="Helpers\HttpClientsFactory.cs" />
@@ -130,6 +132,12 @@
     <Compile Include="Pages\LoginPage.xaml.cs">
       <DependentUpon>LoginPage.xaml</DependentUpon>
     </Compile>
+    <Compile Include="Pages\NewTaskPage.xaml.cs">
+      <DependentUpon>NewTaskPage.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="Pages\TasksPage.xaml.cs">
+      <DependentUpon>TasksPage.xaml</DependentUpon>
+    </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\CustomDialogService.cs" />
     <Compile Include="Services\GoogleApiService.cs" />
@@ -143,8 +151,10 @@
     <Compile Include="Services\UserCredentialService.cs" />
     <Compile Include="Interfaces\IUserCredentialService.cs" />
     <Compile Include="ViewModels\BaseViewModel.cs" />
-    <Compile Include="ViewModels\HomeViewModel.cs" />
+    <Compile Include="ViewModels\NavViewModel.cs" />
     <Compile Include="ViewModels\LoginViewModel.cs" />
+    <Compile Include="ViewModels\NewTaskViewModel.cs" />
+    <Compile Include="ViewModels\TasksViewModel.cs" />
     <Compile Include="ViewModels\ViewModelLocator.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -237,6 +247,14 @@
       <SubType>Designer</SubType>
     </Page>
     <Page Include="Pages\LoginPage.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="Pages\NewTaskPage.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="Pages\TasksPage.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
