+++ MiraiNotes.UWP/Pages/NewTaskPage.xaml
@@ -1,0 +1,70 @@
+﻿<Page
+    x:Class="MiraiNotes.UWP.Pages.NewTaskPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:converters="using:MiraiNotes.UWP.Converters"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:local="using:MiraiNotes.UWP.Pages"
+    xmlns:localExtensions="using:MiraiNotes.UWP.Extensions"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}"
+    DataContext="{Binding NewTask, Source={StaticResource Locator}}"
+    mc:Ignorable="d">
+
+    <Page.Resources>
+        <converters:BooleanToVisibilityConverter x:Key="BoolToVisibilityConverter" />
+    </Page.Resources>
+
+    <Grid x:Name="MainSplitViewPane">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="2*" />
+            <RowDefinition Height="Auto" />
+        </Grid.RowDefinitions>
+        <ProgressRing
+            Grid.RowSpan="4"
+            Width="75"
+            Height="75"
+            HorizontalAlignment="Center"
+            VerticalAlignment="Center"
+            IsActive="True"
+            Visibility="{Binding ShowTaskProgressRing, Converter={StaticResource BoolToVisibilityConverter}}" />
+        <TextBox
+            x:Name="CurrentTaskTitle"
+            Grid.Row="1"
+            Margin="20,20,20,0"
+            localExtensions:FocusExtension.IsFocused="{Binding IsCurrentTaskTitleFocused, Mode=TwoWay}"
+            AcceptsReturn="True"
+            IsSpellCheckEnabled="False"
+            MaxLength="50"
+            SelectionHighlightColor="Green"
+            Text="{Binding CurrentTask.Title, Mode=OneWay}"
+            TextWrapping="NoWrap" />
+        <TextBox
+            Grid.Row="2"
+            Margin="20"
+            IsSpellCheckEnabled="False"
+            PlaceholderText="Enter your task text here"
+            ScrollViewer.VerticalScrollBarVisibility="Auto"
+            Text="{Binding CurrentTask.Notes, Mode=TwoWay}"
+            TextWrapping="Wrap" />
+        <StackPanel
+            Grid.Row="3"
+            Margin="0,0,0,20"
+            HorizontalAlignment="Center"
+            VerticalAlignment="Center"
+            Orientation="Horizontal">
+            <AppBarButton
+                Icon="Save"
+                Label="Save"
+                ToolTipService.ToolTip="Save your task changes" />
+            <AppBarSeparator />
+            <AppBarButton
+                Command="{Binding ClosePaneCommand}"
+                Icon="Cancel"
+                Label="Cancel"
+                ToolTipService.ToolTip="Cancel your task changes" />
+        </StackPanel>
+    </Grid>
+</Page>
